Moved INS to NavigationSystem folder
